# Delegate Int|Short|Long.reverseBytes to Java Stdlib